+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 
 * Add a ``othername`` field on on the ``Person`` model
 * Add a ``letter_prefix`` property on the ``Person`` model
+* Add a ``full_name`` property on the ``Person`` model
 
 0.1.18
 ------